pyproject.toml
[project]
name = "primetrade-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-binance>=1.0.29",
    "python-dotenv>=1.1.1",
]
